+++ oauth_server/src/main/java/bgs/oauth_server/dao/AppsAccessService.java
@@ -30,14 +30,14 @@ public class AppsAccessService {
 
     public List<ClientApp> readAll() throws SQLException {
         final MapSqlParameterSource parameters = new MapSqlParameterSource();
-        final String sql = "select * from client_apps";
+        final String sql = "select * from oauth.client_apps";
         return namedJdbcTemplate.query(sql, parameters, (resultSet, i) -> createClientAppFromResult(resultSet));
     }
 
     public ClientApp readById(Integer id) throws SQLException {
         final MapSqlParameterSource parameters = new MapSqlParameterSource();
         parameters.addValue("client_app_id", id);
-        final String sql = "select * from client_apps where client_app_id = :client_app_id";
+        final String sql = "select * from oauth.client_apps where client_app_id = :client_app_id";
         List<ClientApp> result = namedJdbcTemplate.query(sql, parameters, (resultSet, i) -> createClientAppFromResult(resultSet));
         if (!result.isEmpty()) {
             return result.get(0);
@@ -52,7 +52,7 @@ public class AppsAccessService {
         parameters.addValue("redirect_url", object.getRedirectURL());
         parameters.addValue("age_restriction", object.isAgeRestriction());
         parameters.addValue("user_id", object.getUser().getUserId());
-        final String sql = "insert into client_apps (app_secret, redirect_url, age_restriction, user_id) values (:app_secret, :redirect_url, :age_restriction, :user_id)";
+        final String sql = "insert into oauth.client_apps (app_secret, redirect_url, age_restriction, user_id) values (:app_secret, :redirect_url, :age_restriction, :user_id)";
         namedJdbcTemplate.update(sql, parameters);
         return object;
     }
@@ -64,7 +64,7 @@ public class AppsAccessService {
         parameters.addValue("age_restriction", object.isAgeRestriction());
         parameters.addValue("user_id", object.getUser().getUserId());
         parameters.addValue("client_app_id", object.getClientAppId());
-        final String sql = "update client_apps set app_secret=:app_secret, redirect_url=:redirect_url, age_restriction=:age_restriction, user_id=:user_id where client_app_id=:client_app_id";
+        final String sql = "update oauth.client_apps set app_secret=:app_secret, redirect_url=:redirect_url, age_restriction=:age_restriction, user_id=:user_id where client_app_id=:client_app_id";
         namedJdbcTemplate.update(sql, parameters);
         return object;
     }
@@ -72,7 +72,7 @@ public class AppsAccessService {
     public void remove(ClientApp object) throws SQLException {
         final MapSqlParameterSource parameters = new MapSqlParameterSource();
         parameters.addValue("client_app_id", object.getClientAppId());
-        final String sql = "DELETE FROM client_apps WHERE client_app_id = :client_app_id";
+        final String sql = "DELETE FROM oauth.client_apps WHERE client_app_id = :client_app_id";
         namedJdbcTemplate.update(sql, parameters);
     }
 }
